Refactor core program/compiler module into compiler namespace
