+++ src/core/src/compiler/compiler/constants.rs
