+++ src/core/src/compiler/compiler/sections.rs
