+++ src/core/src/compiler/mod.rs
