+++ src/core/src/compiler/symbol_table.rs
